requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
Pillow>=9.1.0,<9.2
torch==2.0.1
torchvision==0.15.2
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-cluster==1.6.1
torch-spline-conv==1.2.2
pandas==2.2.0
uwsgi>=2.0.19,<2.1